pyproject.toml
[project]
name = "tutorial-spark-ducklake"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.3.2",
    "pyspark>=4.0.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
]
